+++ app/models/cmd/tenant.go
@@ -42,7 +42,7 @@ type SaveVerificationKey struct {
 	Request  NewEmailVerification
 }
 
-//NewEmailVerification is used to register a new email verification process
+//NewEmailVerification is used to define an email verification process
 type NewEmailVerification interface {
 	GetEmail() string
 	GetName() string
